models/hair_classifier/requirements.txt
fastai>=2.7.12
gradio>=3.50.2
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
albumentations==1.3.1
opencv-python>=4.8.0  